Enhance cabin management: add duplicate functionality in CabinRow, implement useEditCabin hook, and refactor CreateCabinForm for improved editing capabilities.
