+++ src/features/cabins/CreateCabinForm.jsx
@@ -4,16 +4,16 @@ import Button from "../../ui/Button";
 import FileInput from "../../ui/FileInput";
 import Textarea from "../../ui/Textarea";
 import { useForm } from "react-hook-form";
-import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useCreateCabin } from "./useCreateCabin";
-import { createEditCabin } from "../../services/apiCabins";
-import { toast } from "react-hot-toast";
+import { useEditCabin } from "./useEditCabin";
+
 import FormRow from "../../ui/FormRow";
 
 function CreateCabinForm({ cabinToEdit = {} }) {
   const { id: editId, ...editValue } = cabinToEdit;
   const { isCreating, createCabin } = useCreateCabin();
-
+  const { isEditing, editCabin } = useEditCabin;
+  const isWorking = isCreating || isEditing;
 
   const isEditSession = Boolean(editId);
   const {
@@ -26,23 +26,9 @@ function CreateCabinForm({ cabinToEdit = {} }) {
     defaultValues: isEditSession ? editValue : {},
   });
 
-  const queryClient = useQueryClient();
-
-  const { isLoading: isEditing, mutate: editCabin } = useMutation({
-    mutationFn: ({ newCabinData, id }) => createEditCabin(newCabinData, id),
-    onSuccess: () => {
-      toast.success("Cabin successfully edited");
-      queryClient.invalidateQueries({ queryKey: ["cabins"] });
-      reset();
-    },
-    onError: (err) => {
-      toast.error(err.message);
-    },
-  });
-  const isWorking = isCreating || isEditing;
   function onSubmit(data) {
     const image = typeof data.image === "string" ? data.image : data.image[0];
-    
+
     if (isEditSession)
       editCabin({ newCabinData: { ...data, image: image }, id: editId });
     else
