+++ src/features/cabins/useCabins.js
@@ -1,0 +1,19 @@
+import { useQuery } from "@tanstack/react-query";
+
+import { getCabins } from "../../services/apiCabins";
+
+export function useCabins() {
+  const {
+    isLoading,
+    data: cabins,
+    error,
+  } = useQuery({
+    queryKey: ["cabins"],
+    queryFn: getCabins,
+  });
+  return {
+    isLoading,
+    cabins,
+    error,
+  };
+}
